Newbox (Bill's PC overhaul)
Newbox is a complete overhaul of Bill's PC. It contains a new graphical UI that allows for easier movement of mons between the Party and PC boxes, without needing to save the game. There are now 16 PC boxes with 20 mons each. The underlying code that supports Bill's PC has been completely reworked to add better functionality.

Co-Authored-By: Rangi <35663410+Rangi42@users.noreply.github.com>
Co-Authored-By: Fredrik Ljungdahl <fredde1994@gmail.com>